pyproject.toml
[tool.poetry]
name = "kickstarterpredictor"
version = "0.1.0"
description = ""
authors = ["Lucas Monsegu <lucas.monsegu@budget-insight.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
discord-py = "1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
